+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 - Added option to skip batch sims if output file already exists
 
+- Added option to overlay pop labels and show avg rates to plotRaster() (issue #111)
+
 - Fixed bug positioning cells with 3d geom
 
 - Fixed bug in sim.clearAll by closing all figures instead of current (issue #168)